seperated react touter styled link as separated component
