+++ src/components/ReactRouterStyledLink/index.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+import { StyledLink } from './style';
+
+const ReactRouterStyledLink = ({ to }) => {
+  return <StyledLink to={to} />;
+};
+
+export default ReactRouterStyledLink;
